// src/types/api/service/models/PageTagDto.ts
export interface PageTagDto {
    /**
     * 
     * @type {number}
     * @memberof PageTagDto
     */
    totalElements?: number;
    /**
     * 
     * @type {number}
     * @memberof PageTagDto
     */
    totalPages?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PageTagDto
     */
    first?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PageTagDto
     */
    last?: boolean;
    /**
     * 
     * @type {number}
     * @memberof PageTagDto
     */
    size?: number;
    /**
     * 
     * @type {Array<TagDto>}
     * @memberof PageTagDto
     */
    content?: Array<TagDto>;
    /**
     * 
     * @type {number}
     * @memberof PageTagDto
     */
    number?: number;
    /**
     * 
     * @type {SortObject}
     * @memberof PageTagDto
     */
    sort?: SortObject;
    /**
     * 
     * @type {number}
     * @memberof PageTagDto
     */
    numberOfElements?: number;
    /**
     * 
     * @type {PageableObject}
     * @memberof PageTagDto
     */
    pageable?: PageableObject;
    /**
     * 
     * @type {boolean}
     * @memberof PageTagDto
     */
    empty?: boolean;
}